requirements.txt
gradio>=4.44.0
Pillow>=10.4.0
requests>=2.32.0
python-dotenv>=1.0.1
openai>=1.0.0
